+++ libcpu/risc-v/SConscript
@@ -18,11 +18,7 @@ elif rtconfig.CPU == "virt64" :
     group = group
 elif rtconfig.CPU == "c906" :
     group = group
-elif rtconfig.CPU == "ch32v1" :
-    group = group
-elif rtconfig.CPU == "ch32v2" :
-    group = group  
-elif rtconfig.CPU == "ch32v3" :
+elif rtconfig.CPU == "ch32" :
     group = group
 elif rtconfig.CPU == "hpmicro":
     group = group
