--- libcpu/risc-v/ch32v3/SConscript
@@ -1,16 +1,0 @@
-# RT-Thread building script for component
-
-from building import *
-
-Import('rtconfig')
-
-cwd     = GetCurrentDir()
-src     = Glob('*.c') + Glob('*.cpp')
-CPPPATH = [cwd]
-
-if rtconfig.PLATFORM == 'gcc':
-    src += Glob('*_gcc.S')
-
-group = DefineGroup('cpu', src, depend = [''], CPPPATH = CPPPATH)
-
-Return('group')
